GameRepositoryとなる部分を別ファイルに分離
